+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/common/fqName.kt
@@ -1,14 +1,9 @@
-package com.javiersc.kotlin.compiler.extensions.ir
+package com.javiersc.kotlin.compiler.extensions.common
 
 import kotlin.reflect.KClass
-import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
-import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.name.FqName
 
 public fun String.toFqName(): FqName = FqName(this)
-
-public val IrFunctionAccessExpression.packageFqName: FqName
-    get() = symbol.owner.kotlinFqName.parent()
 
 public inline fun <reified T> fqName(): FqName = T::class.toFqName()