+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/common/name.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.compiler.extensions.common
+
+import org.jetbrains.kotlin.name.Name
+
+public fun String.toName(): Name = Name.identifier(this)