+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irAnnotationContainer.kt
@@ -1,0 +1,8 @@
+package com.javiersc.kotlin.compiler.extensions.ir
+
+import com.javiersc.kotlin.compiler.extensions.common.fqName
+import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
+import org.jetbrains.kotlin.ir.util.hasAnnotation
+
+public inline fun <reified T : Annotation> IrAnnotationContainer.hasAnnotation(): Boolean =
+    annotations.hasAnnotation(fqName<T>())