+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irClassId.kt
@@ -7,4 +7,5 @@ public fun String.toClassId(): ClassId = FqName(this).run { ClassId(parent(), sh
 
 public fun FqName.toClassId(): ClassId = ClassId(parent(), shortName())
 
-public inline fun <reified T> classId(): ClassId = fqName<T>().toClassId()
+public inline fun <reified T> classId(): ClassId = com.javiersc.kotlin.compiler.extensions.common.fqName<T>()
+    .toClassId()