+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irFunctionAccessExpression.kt
@@ -1,28 +1,28 @@
 package com.javiersc.kotlin.compiler.extensions.ir
 
-import org.jetbrains.kotlin.ir.IrElement
-import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
-import org.jetbrains.kotlin.ir.declarations.IrClass
+import com.javiersc.kotlin.compiler.extensions.common.fqName
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
 import org.jetbrains.kotlin.ir.util.hasAnnotation
+import org.jetbrains.kotlin.ir.util.kotlinFqName
+import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.name.Name
+
+public val IrFunctionAccessExpression.callableId: CallableId
+    get() = CallableId(packageFqName, name)
+
+public val IrFunctionAccessExpression.packageFqName: FqName
+    get() = symbol.owner.kotlinFqName.parent()
+
+public val IrFunctionAccessExpression.name: Name
+    get() = symbol.owner.name
 
 public val IrFunctionAccessExpression.annotations: List<IrConstructorCall>
     get() = symbol.owner.annotations
-
-public inline fun <reified T : Annotation> IrAnnotationContainer.hasAnnotation(): Boolean =
-    annotations.hasAnnotation(fqName<T>())
 
 public inline fun <reified T : Annotation> IrFunctionAccessExpression.hasAnnotation(): Boolean =
     annotations.hasAnnotation(fqName<T>())
 
 public fun IrFunctionAccessExpression.hasAnnotation(annotation: FqName): Boolean =
     annotations.hasAnnotation(annotation)
-
-public fun IrElement.hasAnnotation(annotation: FqName): Boolean =
-    when (this) {
-        is IrClass -> hasAnnotation(annotation)
-        is IrFunctionAccessExpression -> hasAnnotation(annotation)
-        else -> false
-    }