+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irValueParameter.kt
@@ -1,7 +1,5 @@
 package com.javiersc.kotlin.compiler.extensions.ir
 
-import org.jetbrains.kotlin.ir.expressions.IrCall
-
 //fun IrCall.mapValueParameters(
 //    map: (IrValueParameter) -> IrValueParameter,
 //): IrCall = copyValueArguments(valueArguments.map(map))